app/requirements.txt
amqp==2.6.0
asgiref==3.2.10
autopep8==1.5.3
billiard==3.6.3.0
Django==3.0.7
django-appconf==1.0.4
django-filter==2.3.0
django-post-office==3.4.1
django-taggit==1.3.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
drf-extensions==0.6.0
future==0.18.2
gunicorn==20.0.4
jsonfield==3.1.0
kombu==4.6.11
Markdown==3.2.2
psycopg2-binary==2.8.5
pycodestyle==2.6.0
PyJWT==1.7.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
vine==1.3.0
